ensure ChangePrifilePicture not calls deletefile if file does not exists and saveUserPicture Throws

+++ src/domain/use-cases/change-profile-picture.ts
@@ -20,7 +20,11 @@ export const setupChangeProfilePicture: setup = (fileStorage, crypto, userProfil
   try {
     await userProfileRepo.savePicture(userProfile)
   } catch (error) {
-    await fileStorage.delete({ key })
+    if (file !== undefined) {
+      await fileStorage.delete({ key })
+    }
+
+    throw new Error()
   }
 
   return userProfile
